fix: extraction of @elgato/schemas (#28)
* fix: extraction of @elgato/schemas

Previously, completion would occur when the contents of the HTTP stream had been fully read, and not when the contents had been extract. This changes updates the flow so that the file is first downloaded, integrity checked, and then installed. There is also a fallback in place to re-install the previous version should the update fail.

* refactor: switch file name to uuid

---------

Co-authored-by: Richard Herman <geekyeggo@users.noreply.github.com>